refs #8874 - rework POT/PO updates for gettext 3's edit.po
3.1.13 adds an intermediate .edit.po file alongside each .po, which is meant
to be kept outside of SCM and updated by users, whereupon it's merged back into
the .po on the next rake gettext:find execution.

Previously we overwrote all gettext .po file changes with files directly from
Transifex, but now these are downloaded to .edit.po and gettext merges them
back in.

To update a plugin's translations, this now requires the maintainer to:

1. run `make -C locale tx-update` to pull translations to .edit.po
2. run `rake plugin:gettext[foreman_plugin]` (Foreman) to merge .edit.po to .po
3. run `make -C locale mo-files` to refresh MO binaries from PO files