Force pulling in the correct updated translations
When pulling from Transifex what we want is:

* All existing languages
* All new languages with at least 50%

The tx tool can't handle this: it either applies a minimum percentage
and skips anything below it. That ends up clearing all translations for
the languages that dropped below 50%.

We can drop any language that drops below 50%, but in this approach I'm
forcing an update of all existing languages. I've chosen the latter
approach because we don't have code to automatically (properly) drop a
language.